Fix failing to edit entry if not reachable
Calendar fails to edit entry if the address of the calendar isn't reachable (or whatever other reason). This PR allows us to edit the entry anyway so that it can be reused.
Think of the scenario: A user changes the address of his caldav server, then tries to edit the entry in FreePBX but he cannot so he has to recreate everything from scratch (even calendar groups or other things!). This way the user can easily change the parameters and still use the same entry